Moved SecretService to dedicated submodule, added empty windows service class, and updated SecretService imports
